add!: transformation stage for CNTGFuse
Also fixes typo in USAGE.md

CNTGFuse behave similarly to FuseFuzzers. The difference is the following
* CNTGFuse deals with simple API combinations with function signature of
  `int(void)`
* FuseFuzzer deals with fuzz drivers with function signatures of
  `i-dont-remember(char*, size_t)`
* CNTGFuse does not have exploit_transform

Currently, only the transformation stage is done, which takes the seeds
generated and do some sanitization on each seed (such as preventing buffer
overflow). These transformed seeds are stored in the temporary
"output/<lib>/drivers" directory. They are allocated new IDs that remove empty
space between them (for example, if the first seed was id_000008.cc, it is
assigned id_000000.cc instead).

Attempts to run the command will fail with todo! as intended.